refactor(useProducts): useProducts 추상화 했던 것 useFetchProducts로 단일화 및 네이밍 useProducts로 수정
