+++ src/hooks/useProducts.ts
@@ -1,19 +1,81 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
-import { Product } from '../types/product';
+import { fetchProducts } from '../api/products';
+import usePage from './usePage';
+
+import { Category, Order, Product, Sort } from '../types/product';
 
 const useProducts = () => {
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<unknown>(null);
+  const [isLastPage, setIsLastPage] = useState(false);
   const [products, setProducts] = useState<Product[]>([]);
+  const [category, setCategory] = useState<Category>('all');
+  const [sort, setSort] = useState<Sort>({
+    price: 'asc',
+  });
 
-  const addProducts = (newProducts: Product[]) => {
-    setProducts((prevProducts) => [...prevProducts, ...newProducts]);
+  const { page, increasePage, decreasePage, resetPage } = usePage();
+
+  const getProducts = async () => {
+    try {
+      setLoading(true);
+      const data = await fetchProducts(page, category, sort);
+
+      setIsLastPage(data.last);
+      setProducts((prevProducts) => [...prevProducts, ...data.content]);
+    } catch (error) {
+      decreasePage();
+      setError(error);
+      setTimeout(() => setError(null), 3000);
+    } finally {
+      setLoading(false);
+    }
   };
 
-  const resetProducts = () => {
+  useEffect(() => {
+    if (!isLastPage && !error) getProducts();
+  }, [page, category, sort]);
+
+  const reset = () => {
+    resetPage();
     setProducts([]);
   };
 
-  return { products, addProducts, resetProducts };
+  const filterByCategory = (selectedCategory: Category) => {
+    if (selectedCategory !== category) {
+      reset();
+      setCategory(selectedCategory);
+    }
+  };
+
+  const filterBySort = (condition: string, order: Order) => {
+    if (sort.price !== order) {
+      reset();
+      setSort((prevSort) => {
+        return {
+          ...prevSort,
+          [condition]: order,
+        };
+      });
+    }
+  };
+
+  const fetchNextPage = () => {
+    if (!isLastPage && !error) increasePage();
+  };
+
+  return {
+    products,
+    loading,
+    error,
+    page,
+    fetchNextPage,
+    category,
+    filterByCategory,
+    sort,
+    filterBySort,
+  };
 };
 
 export default useProducts;
